Keep given search payload in the body.
Previously the payload would get encoded in the params as well as used in
the body for a search. The payload would also be removed during the
request, causing it to missing when the request is logged or if it is
reperformed.

Also, avoid encoding :load and :wrapper options in search request url.